Bail early when summing with NaN values
Arithmetic with long doubles and NA/NaN/Inf/-Inf is VERY slow. This has a huge performance gain when any NA/NaN values are present. We exit early if we either already have a nan value, or if we are about to get one. This is still slow with Inf/-Inf, but there is a much larger performance penalty to detect these correctly, and is much rarer anyways