Switch to C++11, since we know that is supported
This now uses a single C++11 compilation unit to get access to `alignof()`, which we call on the two types we need it for, `long double` and our custom `struct mean_state_t`. Forward declaring it in `summary-core.h` is enough to give us access to those C++ functions from C. We forward declare rather than including `summary-core-align.hpp`, as that will attempt to compile `extern "C"` with a C compiler, which won't work.